Written the function print_file_from_the_end()
The function print_file_from_the_end() has now been written and tested.
The problem to be fixed is that the programm seems not to run if I use
methods such as .tellg() and .seekg() (both in the function get_cnt_of_lines_in_file()).
So this function is commented at the moment.

Also, a function that checks if the arguments are written correctly has
to be written (function check_cla).